Moving annotations from the reader reducer (#4060)
* refactoring documents code to document reducer

* Updating a bit more code

* refactoring a bit more and updating constants

* Moving couple of actions to a different file

* Removing unused import

* Moving a bit more code

* Fixing lint issues

* Including error information in the reducer

* Removing update

* Moving some reducer code around

* Reusing existing appeal deatails

* Making using combine reducers on documents

* Updating the test

* moving reducer code

* Changing the combinedReducer function

* Modifying the order of actions

* Fixing imports

* Updating variables

* Moving the annotation layer

* Changing merge logic

* fixing actions and fixing tests

* Fixing tests

* moving a reusable method

* Moving back methods

* Refactoring filtering code in the reducer

* Minor method refactor

* Moving constants

* More annotation layer refactor

* Changes to annotation layer

* Changing where annotations are injected

* Fixing some lint issues

* Changing to annotation layer from reader reducer

* flattening the annotation reducer

* Also flattening the pdf in annotation reducer

* Renaming a method name

* Correcting isPlacingAnnotation selectors

* Padding the error message to the reducer
