+++ client/app/reader/CommentIcon.jsx
@@ -55,7 +55,7 @@ CommentIcon.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  ..._.pick(state.readerReducer.ui, 'selectedAnnotationId')
+  ..._.pick(state.annotationLayer, 'selectedAnnotationId')
 });
 
 export default connect(mapStateToProps)(CommentIcon);
