+++ client/app/reader/Pdf.jsx
@@ -11,7 +11,7 @@ import { togglePdfSidebar } from '../reader/PdfViewer/PdfViewerActions';
 import { onScrollToComment } from '../reader/Pdf/PdfActions';
 import { placeAnnotation, startPlacingAnnotation,
   stopPlacingAnnotation, showPlaceAnnotationIcon
-} from '../reader/PdfViewer/AnnotationActions';
+} from '../reader/AnnotationLayer/AnnotationActions';
 
 import { INTERACTION_TYPES, CATEGORIES } from '../reader/analytics';
 
@@ -99,9 +99,10 @@ export class Pdf extends React.PureComponent {
 const mapStateToProps = (state, props) => {
   return {
     ...state.readerReducer.ui.pdf,
-    ..._.pick(state.readerReducer, 'placingAnnotationIconPageCoords'),
+    ..._.pick(state.annotationLayer, 'placingAnnotationIconPageCoords'),
     rotation: _.get(state.readerReducer.documents, [props.documentId, 'rotation']),
-    sidebarHidden: state.readerReducer.ui.pdf.hidePdfSidebar
+    sidebarHidden: state.readerReducer.ui.pdf.hidePdfSidebar,
+    isPlacingAnnotation: state.annotationLayer.isPlacingAnnotation
   };
 };
