+++ client/app/reader/PdfUI.jsx
@@ -12,7 +12,7 @@ import { connect } from 'react-redux';
 import { selectCurrentPdf, resetJumpToPage, togglePdfSidebar, toggleSearchBar
 } from '../reader/PdfViewer/PdfViewerActions';
 import { rotateDocument } from '../reader/Pdf/PdfActions';
-import { stopPlacingAnnotation } from '../reader/PdfViewer/AnnotationActions';
+import { stopPlacingAnnotation } from '../reader/AnnotationLayer/AnnotationActions';
 import { docListIsFiltered } from '../reader/selectors';
 import { DownloadIcon, FilterIcon, PageArrowLeft, PageArrowRight, LeftChevron,
   ExternalLink, FitToScreen, Rotate, SearchIcon } from '../components/RenderFunctions';
@@ -288,6 +288,7 @@ const mapStateToProps = (state, props) => {
     ..._.pick(state.readerReducer.ui, 'filteredDocIds'),
     docListIsFiltered: docListIsFiltered(state.readerReducer),
     loadError: state.readerReducer.documentErrors[props.doc.content_url],
+    isPlacingAnnotation: state.annotationLayer.isPlacingAnnotation,
     ...state.readerReducer.ui.pdf,
     numPages
   };
