+++ client/app/reader/commonActions.js
@@ -1,6 +1,7 @@
 import {
   HIDE_ERROR_MESSAGE,
-  SHOW_ERROR_MESSAGE
+  SHOW_ERROR_MESSAGE,
+  UPDATE_FILTERED_DOC_IDS
 } from './constants';
 
 // errors
@@ -12,9 +13,23 @@ export const hideErrorMessage = (messageType) => ({
   }
 });
 
-export const showErrorMessage = (messageType) => ({
+export const showErrorMessage = (messageType, errorMessage) => ({
   type: SHOW_ERROR_MESSAGE,
   payload: {
-    messageType
+    messageType,
+    errorMessage
   }
 });
+
+// Apply filters
+
+export const updateFilteredIds = () => (dispatch, getState) => {
+  const { annotationLayer } = getState();
+
+  dispatch({
+    type: UPDATE_FILTERED_DOC_IDS,
+    payload: {
+      annotationLayer
+    }
+  });
+};
