+++ client/test/node/reader/DecisionReviewer-test.js
@@ -19,7 +19,9 @@ import { formatDateStr } from '../../../app/util/DateUtil';
 import readerReducer from '../../../app/reader/reducer';
 import caseSelectReducer from '../../../app/reader/CaseSelect/CaseSelectReducer';
 import PdfJsStub, { PAGE_WIDTH, PAGE_HEIGHT } from '../../helpers/PdfJsStub';
-import { onReceiveDocs, onReceiveAnnotations } from '../../../app/reader/LoadingScreen/LoadingScreenActions';
+import { onReceiveDocs } from '../../../app/reader/LoadingScreen/LoadingScreenActions';
+import { onReceiveAnnotations } from '../../../app/reader/AnnotationLayer/AnnotationActions';
+
 import sinon from 'sinon';
 import { AutoSizer } from 'react-virtualized';
 const vacolsId = 'reader_id1';
